[#64]feat: 가이드&여행자의 세션 룸 페이지 통합
